% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{horusheresy}
\alias{horusheresy}
\title{Data on Horus Heresy}
\format{
A data frame with 390 rows and 2 columns.
\describe{
 \item{name}{Name of the rule}
 \item{text}{Description of the rule}
}
}
\source{
Obtained from the Horus Heresy BattleScribe data repo: https://github.com/BSData/horus-heresy-2nd-edition
}
\usage{
horusheresy
}
\description{
A dataset containing information on the Horus Heresy 2.0 rules
}
\examples{
data(horusheresy)
}
\keyword{datasets}
